requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.7
pydantic==2.5.3
redis==5.0.1
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
python-dotenv==1.0.0